// src/app/models/mundo.model.ts
export interface Mundo {
    id: string
    nombre: string
    detalles: string
    imagen: string

    universoId: string

    diferente: string
    funciona: string
    reglas: string
    geografia: string
    historia: string
    eventEspeciales: string

    updatedAt: number;
    _deleted: boolean;
}